indexing_function/requirements.txt
google-cloud-firestore
faiss-cpu
google-cloud-storage
sentence-transformers